# util/httputil: reduces heap allocations in newCompressedResponseWriter