`newCompressedResponseWriter` currently calls [strings.Split](https://pkg.go.dev/strings#Split) on the value of request's _Accept-Encoding_ header, which is untrusted data. As a result, in the face of a malicious request whose _Accept-Encoding_ header is full of commas, the handler in question allocates O(n) bytes (where n stands for the length of the request's _Accept-Encoding_ value), with a constant factor of about 16. Relevant weakness: [CWE-405: Asymmetric Resource Consumption (Amplification)](https://cwe.mitre.org/data/definitions/405.html)

With this change, `newCompressedResponseWriter` allocates O(1) bytes even in the face of such malicious requests.

---

Running the benchmark:

```shell
git checkout 827f47f466aad632e7037c0b6d52099fa8a7f33a
go test -benchmem -run=^$ -count 10 -bench '^BenchmarkNewCompressionHandler_MaliciousAcceptEncoding$' github.com/prometheus/prometheus/util/httputil > new
git checkout f72a58a218161bd1061c1ccc7fae513616d4a75a
go test -benchmem -run=^$ -count 10 -bench '^BenchmarkNewCompressionHandler_MaliciousAcceptEncoding$' github.com/prometheus/prometheus/util/httputil > old
benchstat old new
```

Some results:

```txt
goos: darwin
goarch: amd64
pkg: github.com/prometheus/prometheus/util/httputil
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
                                                │     old     │                 new                 │
                                                │   sec/op    │   sec/op     vs base                │
NewCompressionHandler_MaliciousAcceptEncoding-8   18.60m ± 2%   13.54m ± 3%  -27.17% (p=0.000 n=10)

                                                │       old        │                 new                 │
                                                │       B/op       │    B/op     vs base                 │
NewCompressionHandler_MaliciousAcceptEncoding-8   16785442.50 ± 0%   32.00 ± 0%  -100.00% (p=0.000 n=10)

                                                │    old     │                new                 │
                                                │ allocs/op  │ allocs/op   vs base                │
NewCompressionHandler_MaliciousAcceptEncoding-8   2.000 ± 0%   1.000 ± 0%  -50.00% (p=0.000 n=10)

```