@jub0bs 2025-02-11
@machine424

> Nice catch!

Thanks!

> There is [golang/go@a1ea78c](https://github.com/golang/go/commit/a1ea78c470d3136b7aed42a4d8b94497563f98ea) (replacing `Seq` iterators via `SplitSeq` in `net/http`), but I think your `Cut` approach should be more efficient (didn't run any benchmarks).

I considered waiting until the advent of `strings.SplitSeq` (I'm already a big fan of it), but

- that would require bumping the project's `go` directive (currently at `1.22.7`) to `1.24`, and I'm not sure how hard that would be or how that would impact the rest of the project;
- I indeed expect `strings.Cut` to be more efficient: contrary to `strings.SplitSeq`, it doesn't cause any function literal (the resulting `iter.Seq[string]`) to escape to the heap.

> If the performance is similar, maybe we can just stick to `SplitSeq` as it's more straightforward, otherwise, I think you can convince go folks to move to `Cut` as well.

How about using `strings.Cut` for now and revisiting this decision when the project updates to Go 1.24? I can make a note to come back to this then.